documented source-data-extractor.js and util files.
