+++ lib/services/semantic-form-management.js
@@ -9,7 +9,7 @@ import { FORM, Graph, parse, RDF, RDFNode } from '../util/rdflib';
 import { validateForm } from '@lblod/submission-form-helpers';
 import { SemanticFormBundle } from '../entities/semantic-form-bundle';
 import { META_FILE_MATCHER } from './meta-data';
-import { SourceDataGeneration } from './source-data-generation';
+import { SourceDataExtractor } from './source-data-extractor';
 import { CONFIG_FILE_MATCHER, SPEC_FILE_MATCHER } from './configuration';
 
 /**
@@ -20,7 +20,7 @@ export class SemanticFormManagement {
   constructor(config_files, meta) {
     this.config_files = config_files;
     this.meta = meta;
-    this.source = new SourceDataGeneration(this.config_files);
+    this.source = new SourceDataExtractor(this.config_files);
     this.mutationService = new NTriplesMutation({sudo: true});
   }
 
@@ -91,7 +91,7 @@ export class SemanticFormManagement {
      * NOTE: we expect that at this point sources have been set on the semantic-form
      */
     const bundle = new SemanticFormBundle();
-    bundle.source = await this.source.generate(form.uri, form.getSource(CONFIG_FILE_MATCHER).content['resource']);
+    bundle.source = await this.source.extract(form.uri, form.getSource(CONFIG_FILE_MATCHER).content['resource']);
     bundle.specification = form.getSource(SPEC_FILE_MATCHER);
     // TODO: add some post-processing on the meta-data to enforce business logic.
     bundle.meta = form.getSource(META_FILE_MATCHER);
@@ -205,8 +205,8 @@ function isValid(form, {source, specification}) {
     sourceNode: RDFNode(form.uri),
     store,
   };
-  parse(specification, store, {graph: options.formGraph});
-  parse(source, store, {graph: options.sourceGraph});
+  parse(specification.content, store, {graph: options.formGraph});
+  parse(source.content, store, {graph: options.sourceGraph});
   const form_statements = store.any(undefined, RDF('type'), FORM('Form'), RDFNode('http://form-graph'));
   return validateForm(form_statements, options);
 }
