+++ lib/util/database.js
@@ -19,6 +19,12 @@ export async function update(q, sudo = false) {
   return sudo ? await updateSudo(q) : await muUpdate(q);
 }
 
+/**
+ * Made for the edge case where you want te ?graph, but ensure you are auth.
+ *
+ * @param q - query
+ * @returns {Promise<null|*|*>}
+ */
 export async function authenticatedQuery(q) {
   const response = await query(q);
   // NOTE: if nothing was found we could assume we did not have the rights
