+++ lib/util/file.js
@@ -5,7 +5,12 @@ import { resolve } from 'path';
  * JS file containing all helpers for working with files
  */
 
-
+/**
+ * Returns all files with a directories and it's sub-directories
+ *
+ * @param dir - root directory
+ * @returns {Promise<string[]>}
+ */
 export async function getFiles(dir) {
   const dirents = await fs.readdir(dir, {withFileTypes: true});
   const files = await Promise.all(dirents.map((dirent) => {
@@ -13,10 +18,6 @@ export async function getFiles(dir) {
     return dirent.isDirectory() ? getFiles(res) : res;
   }));
   return Array.prototype.concat(...files);
-}
-
-export function uriToPath(uri) {
-  return uri.replace('data://', '/data/').replace('config://', '/config/');
 }
 
 export function uriToFilename(uri) {
