+++ lib/util/rdflib.js
@@ -4,7 +4,7 @@ import { graph as Graph, parse as rdflibParse, serialize as rdflibSerialize, sym
  * JS file containing all helpers for working with RDFlib
  */
 
-// name-spaces used to query with rdflib.
+// NOTE: name-spaces used to query with rdflib.
 export const RDF = Namespace('http://www.w3.org/1999/02/22-rdf-syntax-ns#');
 export const FORM = Namespace('http://lblod.data.gift/vocabularies/forms/');
